+++ device/esp_tinyusb/tinyusb.c
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+#include <string.h>
 #include "sdkconfig.h"
 #include "esp_log.h"
 #include "esp_check.h"